refactor!: packages in pyproject
Remove environments from package, rename agents/ to autoagora-agents/.
