+++ autoagora_agents/reinforcement_learning_policy_mixins.py
@@ -3,7 +3,7 @@
 
 import torch
 
-from agents.policy_mixins import ExperienceBufferPolicy
+from autoagora_agents.policy_mixins import ExperienceBufferPolicy
 
 
 class VanillaPolicyGradient(ExperienceBufferPolicy):
